refactor: 서비스 분리 및 리팩터링
