+++ server/src/api/auth/services/index.ts
@@ -1,0 +1,10 @@
+export {
+  getUserVerifiesIdByEmail,
+  getUserVerifiesIdByPhone,
+} from './credentials';
+export { default as createUser } from './sign-in';
+export {
+  sendSecurityCodeByRules,
+  sendSecurityCodeEmail,
+  sendSecurityCodeSms,
+} from './sms';
